gcp-pro-dev-votes-function/requirements.txt
functions-framework==3.*
google-cloud-spanner==3.35.*